refactor: rename withSchema.ts → defineResource.ts; drop spike/RFC artifacts
- rename resources/withSchema.ts → resources/defineResource.ts (defineResource is
  the primary API; Resource.withSchema stays the class-form name) + the test file
- remove docs/rfcs/ (the *-real.check.ts type-contract proofs + tsconfig) — a real
  PR shouldn't carry spike/RFC scaffolding; those live in the design PR (#1503)
- strip references to the spikes and the RFC doc (which are not in this PR) from
  code/test comments and resources/DESIGN.md; keep the #1503 pointer for the record

No behavior change. 100 unit tests green.
