--- docs/rfcs/spikes/0001/tsconfig.json
@@ -1,8 +1,0 @@
-{
-	"extends": "../../../../tsconfig.json",
-	"compilerOptions": {
-		"noEmit": true,
-		"strict": true
-	},
-	"include": ["./**/*.ts"]
-}
